Reorganise principle data classes into .core package
