+++ app/src/main/java/com/w3e/sharingmate/core/ExpenseItem.java
@@ -1,4 +1,4 @@
-package com.w3e.sharingmate;
+package com.w3e.sharingmate.core;
 
 import java.util.ArrayList;
 import java.util.Calendar;
